+++ common-module/src/main/java/org/c4marathon/assignment/enums/AccountPolicy.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.domain.model.account.enums;
+package org.c4marathon.assignment.enums;
 
 import lombok.Getter;
