+++ common-module/src/main/java/org/c4marathon/assignment/enums/AccountType.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.domain.model.account.enums;
+package org.c4marathon.assignment.enums;
 
 public enum AccountType {
 	MAIN_ACCOUNT,
