+++ common-module/src/main/java/org/c4marathon/assignment/enums/SavingType.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.domain.model.account.enums;
+package org.c4marathon.assignment.enums;
 
 public enum SavingType {
 	FIXED, // 정기 적금
