+++ common-module/src/main/java/org/c4marathon/assignment/enums/TransferType.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.domain.model.transfer.enums;
+package org.c4marathon.assignment.enums;
 
 public enum TransferType {
 	IMMEDIATE,
