+++ common-module/src/main/java/org/c4marathon/assignment/model/policy/SettlementPolicy.java
@@ -1,0 +1,5 @@
+package org.c4marathon.assignment.model.policy;
+
+public enum SettlementPolicy {
+	EQUAL, RANDOM
+}
